requirements.txt
langchain
python-dotenv
langchain-community
faiss-cpu
langchain-text-splitters
langchain-openai
pandas
streamlit
validators
requests

